Conflicts:
	app/BUILD.bazel
	app/src/main/java/org/oppia/android/app/testing/ConceptCardFragmentTestActivity.kt
	app/src/main/java/org/oppia/android/app/topic/conceptcard/ConceptCardFragment.kt
	app/src/sharedTest/java/org/oppia/android/app/databinding/BUILD.bazel
	app/src/sharedTest/java/org/oppia/android/app/splash/BUILD.bazel
	app/src/test/java/org/oppia/android/app/activity/route/BUILD.bazel

+++ scripts/src/java/org/oppia/android/scripts/apkstats/ComputeAabDifferences.kt
@@ -181,14 +181,14 @@ class ComputeAabDifferences(
   ): FileSizeStats {
     println("Computing file size for: $apkWithoutChangesPath and $apkWithChangesPath")
     val (fileSizeWithoutChanges, downloadSizeWithoutChanges) = if (apkWithoutChangesPath != null) {
-      apkAnalyzerClient.computeFileSize(apkWithoutChangesPath).toLong() to
-        apkAnalyzerClient.computeDownloadSize(apkWithoutChangesPath).toLong()
+      apkAnalyzerClient.computeFileSize(apkWithoutChangesPath) to
+        apkAnalyzerClient.computeDownloadSize(apkWithoutChangesPath)
     } else 0L to 0L
 
     println("Computing estimated download size for: $apkWithoutChangesPath and $apkWithChangesPath")
     val (fileSizeWithChanges, downloadSizeWithChanges) = if (apkWithChangesPath != null) {
-      apkAnalyzerClient.computeFileSize(apkWithChangesPath).toLong() to
-        apkAnalyzerClient.computeDownloadSize(apkWithChangesPath).toLong()
+      apkAnalyzerClient.computeFileSize(apkWithChangesPath) to
+        apkAnalyzerClient.computeDownloadSize(apkWithChangesPath)
     } else 0L to 0L
 
     return FileSizeStats(
